Refactor neo4j query function calls
